Users management in admin section

+++ src/utils/api/resources/user.ts
@@ -7,6 +7,7 @@ import client from "src/utils/api/axios";
 import { IFolder } from "./folder";
 import { IFile } from "./file";
 import { handleAxiosError } from "src/utils/helpers";
+import { EmptyResponse } from "../logic";
 
 export type IUser = {
   id: string;
@@ -16,15 +17,25 @@ export type IUser = {
   isAdmin: boolean;
 };
 
-type MeResponse = {
+type AxiosMeResponse = {
   user: IUser;
   folders: IFolder[];
   files: IFile[];
 };
 
-type UserResponse = {
+type AxiosUserResponse = {
   user: IUser;
 };
+
+type AxiosUsersResponse = {
+  users: IUser[];
+};
+
+type MeResponse = FailureOrSuccess<IError, AxiosMeResponse>;
+
+type UserResponse = FailureOrSuccess<IError, AxiosUserResponse>;
+
+type UsersResponse = FailureOrSuccess<IError, IUser[]>;
 
 type UpdateUserParams = {
   username?: string;
@@ -34,21 +45,19 @@ type UpdateUserParams = {
 export class User {
   private static prefix = "/user" as const;
 
-  static me = async (): Promise<FailureOrSuccess<IError, MeResponse>> => {
+  static me = async (): Promise<MeResponse> => {
     try {
-      const response = await client.get<MeResponse>(`${this.prefix}/me`);
+      const response = await client.get<AxiosMeResponse>(`${this.prefix}/me`);
       return success(response.data);
     } catch (error) {
       return handleAxiosError(error);
     }
   };
 
-  static update = async (
-    params: UpdateUserParams
-  ): Promise<FailureOrSuccess<IError, UserResponse>> => {
+  static updateMe = async (params: UpdateUserParams): Promise<UserResponse> => {
     try {
-      const response = await client.patch<UserResponse>(
-        `${this.prefix}/`,
+      const response = await client.patch<AxiosUserResponse>(
+        `${this.prefix}/me`,
         params
       );
 
@@ -58,9 +67,45 @@ export class User {
     }
   };
 
-  static delete = async (): Promise<FailureOrSuccess<IError, null>> => {
+  static deleteMe = async (): Promise<EmptyResponse> => {
     try {
-      await client.delete(`${this.prefix}/`);
+      await client.delete(`${this.prefix}/me`);
+      return success(null);
+    } catch (error) {
+      return handleAxiosError(error);
+    }
+  };
+
+  static getAll = async (): Promise<UsersResponse> => {
+    try {
+      const response = await client.get<AxiosUsersResponse>(this.prefix);
+
+      return success(response.data.users);
+    } catch (error) {
+      return handleAxiosError(error);
+    }
+  };
+
+  static adminUpdate = async (
+    userId: string,
+    user: Partial<Omit<IUser, "id">>
+  ) => {
+    try {
+      const response = await client.patch<AxiosUserResponse>(
+        `${this.prefix}/${userId}`,
+        user
+      );
+
+      return success(response.data);
+    } catch (error) {
+      return handleAxiosError(error);
+    }
+  };
+
+  static adminDelete = async (userId: string): Promise<EmptyResponse> => {
+    try {
+      await client.delete(`${this.prefix}/${userId}`);
+
       return success(null);
     } catch (error) {
       return handleAxiosError(error);
